Users & Admin Routes [Course section 11]

+++ controllers/jobsController.js
@@ -3,6 +3,8 @@ const geocoder = require('../utils/geocoder');
 const ErrorHandler = require('../utils/errorHandler');
 const catchAsyncErrors = require('../middlewares/catchAsyncErrors');
 const APIFilters = require('../utils/apiFilters');
+const path = require('path');
+const fs = require('fs');
 
 // Get all Jobs => /api/v1/jobs
 exports.getJobs = async (req, res, next) => {
@@ -25,7 +27,10 @@ exports.getJobs = async (req, res, next) => {
 
 // Get a Job by id and slug => /api/v1/job/:id/:slug
 exports.getJob = catchAsyncErrors(async (req, res, next) => {
-    const job = await Job.find({ $and: [{ _id: req.params.id }, { slug: req.params.slug }] });
+    const job = await Job.find({ $and: [{ _id: req.params.id }, { slug: req.params.slug }] }).populate({
+        path: 'user',
+        select: 'name'
+    });
 
     if (!job || job.length === 0) {
         return next(new ErrorHandler('Job not found!', 404));
@@ -58,6 +63,10 @@ exports.updateJob = catchAsyncErrors(async (req, res, next) => {
         return next(new ErrorHandler('Job not found!', 404));
     }
 
+    if (job.user.toString() != req.user.id && req.user.role != 'admin') {
+        return next(new ErrorHandler('User is not allowed to update this job!', 403));
+    }
+
     job = await Job.findByIdAndUpdate(req.params.id, req.body, {
         new: true,
         runValidators: true
@@ -72,11 +81,25 @@ exports.updateJob = catchAsyncErrors(async (req, res, next) => {
 
 // Delete a Job => /api/v1/job/:id
 exports.deleteJob = catchAsyncErrors(async (req, res, next) => {
-    let job = await Job.findById(req.params.id);
+    let job = await Job.findById(req.params.id).select('+candidates');
 
     if (!job) {
         return next(new ErrorHandler('Job not found!', 404));
     }
+
+    if (job.user.toString() != req.user.id && req.user.role != 'admin') {
+        return next(new ErrorHandler('User is not allowed to delete this job!', 403));
+    }
+
+    job.candidates.forEach(candidate => {
+        let filePath = `${__dirname}/public/uploads/${candidate.resume}`.replace('\\controllers', '');
+
+        fs.unlink(filePath, err => {
+            if (err) {
+                return console.log(err);
+            }
+        });
+    })
 
     job = await Job.findByIdAndDelete(req.params.id);
 
@@ -135,4 +158,64 @@ exports.jobStats = catchAsyncErrors(async (req, res, next) => {
         success: true,
         data: stats
     });
+});
+
+// Apply to a Job => /api/v1/job/:id/apply
+exports.applyJob = catchAsyncErrors(async (req, res, next) => {
+    let job = await Job.findById(req.params.id).select('+candidates');
+
+    if (!job) {
+        return next(new ErrorHandler('Job not found!', 404));
+    }
+
+    if (job.lastDate < new Date(Date.now())) {
+        return next(new ErrorHandler('You cannot apply to this job anymore.', 400));
+    }
+
+    job.candidates.forEach(candidate => {
+        if (candidate.id === req.user.id) {
+            return next(new ErrorHandler('You have already applied to this job!', 400));
+        }
+    });
+
+    if (!req.files) {
+        return next(new ErrorHandler('Please upload your resume!', 400));
+    }
+
+    const file = req.files.file;
+
+    const supportedFiles = /.docx|.pdf/;
+    if (!supportedFiles.test(path.extname(file.name))) {
+        return next(new ErrorHandler('Please upload a .docx or .pdf file!', 400));
+    }
+
+    if (file.size > process.env.MAX_FILE_SIZE) {
+        return next(new ErrorHandler('Please upload a smaller file!', 400));
+    }
+
+    file.name = `${req.user.name.replace(' ', '_')}_${job._id}${path.parse(file.name).ext}`;
+    file.mv(`${process.env.UPLOAD_PATH}/${file.name}`, async err => {
+        if (err) {
+            console.log(err);
+        }
+
+        return next(new ErrorHandler('Failed to upload the Resume', 500));
+    });
+
+    await Job.findByIdAndUpdate(req.params.id, {
+        $push: {
+            candidates: {
+                id: req.user.id,
+                resume: file.name
+            }
+        }
+    }, {
+        new: true
+    });
+
+    res.status(200).json({
+        success: true,
+        message: 'Application completed successfully!',
+        data: file.name
+    })
 });
